完成测试一版，尚有瑕疵

+++ src/com/sgaop/project/module/NutzBootMakerChooseStep.java
@@ -5,6 +5,7 @@ import com.intellij.ide.util.projectWizard.ModuleWizardStep;
 import com.intellij.ide.util.projectWizard.WizardContext;
 import com.intellij.ide.wizard.CommitStepException;
 import com.intellij.openapi.options.ConfigurationException;
+import com.intellij.openapi.ui.Messages;
 import com.sgaop.project.FileUtil;
 import com.sgaop.project.ui.ModuleWizardStepUI;
 import org.apache.commons.io.IOUtils;
@@ -26,12 +27,13 @@ import java.util.HashMap;
  */
 public class NutzBootMakerChooseStep extends ModuleWizardStep {
 
-
     protected final WizardContext wizardContext;
 
     private ModuleWizardStepUI moduleWizardStepUI = new ModuleWizardStepUI();
 
     private String downLoadKey;
+
+    private boolean loadingCompleted = false;
 
     private Gson gson = new Gson();
 
@@ -40,29 +42,24 @@ public class NutzBootMakerChooseStep extends ModuleWizardStep {
     }
 
     @Override
-    public JComponent getComponent() {
-        return moduleWizardStepUI.getRoot();
-    }
-
-    @Override
     public void onWizardFinished() throws CommitStepException {
-        try {
-            HttpClient httpclient = HttpClients.createDefault();
-
-            HttpGet httppost = new HttpGet(moduleWizardStepUI.getMakerUrl().getText() + "/maker/download/" + downLoadKey);
-            String path = this.wizardContext.getProjectFileDirectory();
-            if (this.wizardContext.getProject() != null) {
-                NutzBootModuleBuilder moduleBuilder = (NutzBootModuleBuilder) this.wizardContext.getProjectBuilder();
-                path = moduleBuilder.getPath();
+        if (isNutzBootModuleBuilder()) {
+            try {
+                HttpClient httpclient = HttpClients.createDefault();
+                HttpGet httppost = new HttpGet(moduleWizardStepUI.getMakerUrl().getText() + "/maker/download/" + downLoadKey);
+                String path = this.wizardContext.getProjectFileDirectory();
+                if (this.wizardContext.getProject() != null) {
+                    path = ((NutzBootModuleBuilder) this.wizardContext.getProjectBuilder()).getContentEntryPath();
+                }
+                HttpResponse response = httpclient.execute(httppost);
+                File zipFile = Paths.get(path, "temp.zip").toFile();
+                File dir = Paths.get(path).toFile();
+                FileUtil.writeFile(zipFile, response.getEntity().getContent());
+                FileUtil.extractZipFile(zipFile, dir);
+                zipFile.delete();
+            } catch (Exception e) {
+                throw new CommitStepException(e.getMessage());
             }
-            HttpResponse response = httpclient.execute(httppost);
-            File zipFile = Paths.get(path, "temp.zip").toFile();
-            File dir = Paths.get(path).toFile();
-            FileUtil.writeFile(zipFile, response.getEntity().getContent());
-            FileUtil.extractZipFile(zipFile, dir);
-            zipFile.delete();
-        } catch (Exception e) {
-            throw new CommitStepException(e.getMessage());
         }
     }
 
@@ -83,7 +80,7 @@ public class NutzBootMakerChooseStep extends ModuleWizardStep {
                 return true;
             } else {
                 downLoadKey = null;
-                JOptionPane.showMessageDialog(null, redata.getOrDefault("msg", "服务暂不可用！请稍后再试！"), "错误提示", JOptionPane.ERROR_MESSAGE, null);
+                Messages.showErrorDialog(redata.getOrDefault("msg", "服务暂不可用！请稍后再试！").toString(), "错误提示");
                 return false;
             }
         } catch (Exception e) {
@@ -91,37 +88,33 @@ public class NutzBootMakerChooseStep extends ModuleWizardStep {
         }
     }
 
-    @Override
-    public void updateDataModel() {
-        System.out.println("updateDataModel");
-    }
-
-    @Override
-    public void onStepLeaving() {
-        System.out.println("onStepLeaving");
-    }
-
-    @Override
-    public boolean isStepVisible() {
-        System.out.println("isStepVisible");
-        return true;
-    }
-
-    @Override
-    public void disposeUIResources() {
-        System.out.println("disposeUIResources");
-    }
 
     @Override
     public void updateStep() {
-        try {
-            HttpClient httpclient = HttpClients.createDefault();
-            HttpGet httppost = new HttpGet(moduleWizardStepUI.getMakerUrl().getText() + "/maker.json");
-            HttpResponse response = httpclient.execute(httppost);
-            String json = IOUtils.toString(response.getEntity().getContent(), "UTF-8");
-            moduleWizardStepUI.refresh(json);
-        } catch (Exception e) {
-            JOptionPane.showMessageDialog(null, "网络异常！请稍候尝试", "错误提示", JOptionPane.ERROR_MESSAGE, null);
+        if (!loadingCompleted) {
+            try {
+                HttpClient httpclient = HttpClients.createDefault();
+                HttpGet httppost = new HttpGet(moduleWizardStepUI.getMakerUrl().getText() + "/maker.json");
+                HttpResponse response = httpclient.execute(httppost);
+                String json = IOUtils.toString(response.getEntity().getContent(), "UTF-8");
+                moduleWizardStepUI.refresh(json);
+                loadingCompleted = true;
+            } catch (Exception e) {
+                Messages.showErrorDialog("网络异常！请稍候尝试!" + e.getMessage(), "错误提示");
+            }
         }
+    }
+
+    @Override
+    public void updateDataModel() {
+    }
+
+    @Override
+    public JComponent getComponent() {
+        return moduleWizardStepUI.getRoot();
+    }
+
+    private boolean isNutzBootModuleBuilder() {
+        return this.wizardContext.getProjectBuilder() instanceof NutzBootModuleBuilder;
     }
 }
